Change all "python" interp directives to "python2"
- Currently PTU's python sripts invoke python as #!/usr/bin/python and
assume that python2 is the default python version installed on the
machine.  This assumption fails and PTU does not work on some modern
linux distributions that use python3 as the default, and the problem will only
get worse as more and more distributions migrate to python3.  A clean long-term
fix, per PEP 394, is to use a python2 interpreter directive line in all PTU
scripts.  Since any distribution can have Python 3 living alongside Python2,
this fix is future-proof.